Fixed horizontal problems, renamed file
So, I haven't committed anything for a few days now because I was having some problems with my system.  I wanted it to make a little 3x3 tile for everything, so I would be able to have a little square in the center to show what color the piece was and what kind of piece it was while still showing the tile color it was on.  I messed around with the for loops for a bit, and they finally worked!  I also renamed the .cs file to Board.cs instead of EmptyClass.cs, because that was honestly a pretty stupid name I had just never gotten around to changing anyway.  Next, I want to be able to set up a board, possibly, with pieces for both sides.  I'll have to figure out what each piece should be, but I think there are enough Unicode characters for me to figure out what I want.