Added hold at the end of the generation
I wanted to add a hold so I could see if the actual .exe would work, and it does, which means mono is actually working besides in its own little command prompt window.  yay!  I wonder if I could make a little .exe that just runs that program, cds its way to that folder, uses mcs on my .cs, and leaves the exe for me to use.  Interesting thoughts for the future.